Controllers were moved to controllers/v1 directory
